Switching to my own branch and updating to Quartus II 15.0.1
At the time of this commit, there is still an open pull request on the master branch. To prevent this from getting even longer (and reducing the chance of it ever getting integrated into the Parallax repo even further), I won't make any further changes to master until further notice, except in cases of obvious bugs.
I've updated to the latest version of Quartus because it's needed for the BeMicroCV-A9 and the future Parallax 1-2-3 board which is under development at this time.