BeMicroCV-A9 working!
I want to create a top-level module for each target that contains target-specific features. For example, I want to make it possible to use the DIP switches to select various different connections to pin headers and peripherals.

In preparation for this, I created the new BeMicroCV-A9 target WITHOUT the io[31:0] ports in the .qsf file. Instead, I assigned the signal names from the Technical Reference (http://www.alterawiki.com/uploads/f/f3/Hardware_Reference_Guide_BeMicro_CV_A9_1.2.pdf) and I set the top level module for the new BeMicroCV-A9 "revision" to BeMicroCV-A9.v and used the "bemicrocva9" module as top level entity.

I tried to just create a wire[31:0] to connect the various pins on J1 to the "io" port of the old top module but I couldn't get this to work without adding unnecessary bus drivers and/or generating nonsense errors and/or getting into a situation where Quartus was getting confused and taking forever to fit the design.

The solution was to create a modified top-level module "p1v" which is mostly identical to the old top.v module, but instead of exporting the Propeller pins as a 32-bit inout port, p1v makes the input lines, output lines and direction lines available without combining them. The top level module does the combining and according to the RTL schematics, it does what it's supposed to do.

I also took the Altera-specific PLL-based main clock generator out of the p1v module and moved it into a separate module. This will help in the future when we port the P1V design to the Xilinx (again).

The instructions to download to the BeMicroCV-A9 are roughly identical to the BeMicroCV
- Load the .qpf file into Quartus II Web Edition (free) version 15 or later (earlier versions don't support the Cyclone V with the ARM A9 on board in the free version)
- Select the BeMicroCV-A9 revision
- Build it by clicking the "play" button in the toolbar ("Start Compilation")
- Use the File > Convert Programming File, load the BeMicroCV-A9.cof file and click Generate
- Connect the board to a USB port and run the Programmer. Add the BeMicroCV-A9.jic file generated by the converter in the previous step and select the "program/configure" checkbox, then click Start.

Connect a Prop plug to J1 pins 30 (VSS), 32 (NRES), 34 (TX to Prop), 36 (RX from prop). You can use the normal Propeller tools to compile and download programs, but for now you can only download to RAM. The BeMicroCV-A9 has an EEPROM on board but just like on the BeMicroCV and other targets, it's only 2K.

I will update the documentation in the next commit.