A few minor changes to MindRobots' fpga123 target
- Some minor (layout-only) changes to the top level Verilog file
- Renamed the revision from fpga123 to fpga123-A7 to prevent future confusion when the fpga123 will be a Cyclone V-A9 device.

I didn't rename the fpga123.v file because it will probably be used by both the A7 and the A9 version.

In case you don't know: the 1-2-3 board (here fpga123) is a low-volume item that Parallax makes available to engineers who want to beta-test the Propeller-2. The original plan was to use a Cyclone-5 A7 but after the BeMicroCV-A9 came out, they did some research and decided it would be trivial to put an A9 on the 1-2-3 board too. The A9 version is not ready at this time, but some A7 boards already got produced. Both versions will be supported by the Propeller-2 project and also by us, the P1V "team".